Fix logging: rename logging to logger; handle verbosity correctly (#10)
